feat: add embed block + clean the code
